pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contrabin"
version = "0.2.0"
description = "Contrastive pre-training of binary code representations with simplex interpolation"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Yifan Zhang", email = "yifan.zhang.2@vanderbilt.edu" },
]
keywords = [
    "binary-code",
    "contrastive-learning",
    "representation-learning",
    "reverse-engineering",
    "program-analysis",
    "llvm-ir",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "tokenizers>=0.15",
    "datasets>=2.18",
    "accelerate>=0.28",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "sacrebleu>=2.4",
    "nltk>=3.8",
    "rouge-score>=0.1.2",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "typer>=0.12",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "types-pyyaml",
]
viz = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "umap-learn>=0.5",
]
llm = [
    "openai>=1.20",
]
experiment = [
    "wandb>=0.16",
    "tensorboard>=2.16",
]

[project.urls]
Homepage = "https://github.com/CoderDoge1108/ContraBin"
Paper = "https://arxiv.org/abs/2210.05102"
Zenodo = "https://zenodo.org/records/15219264"
Issues = "https://github.com/CoderDoge1108/ContraBin/issues"

[project.scripts]
contrabin = "contrabin.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["contrabin*"]
exclude = ["tests*", "notebooks*", "scripts*"]

[tool.setuptools.package-data]
contrabin = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["notebooks", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501", "B008", "N806", "N803", "N812"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N802", "N803"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "slow: tests that are slow to run",
    "gpu: tests that require a GPU",
    "network: tests that require network access",
]
